+++ ts_agl/arena/__init__.py
@@ -1,6 +1,7 @@
 from ts_agl.arena.cross_domain_arena import CrossDomainArena, run_cross_domain_arena
 
-__all__ = ["CrossDomainArena", "run_cross_domain_arena", "SafeWriteArena", "run_safe_write_arena", "InteractiveWorkflowArena", "run_interactive_workflow_arena"]
+__all__ = ["CrossDomainArena", "run_cross_domain_arena", "SafeWriteArena", "run_safe_write_arena", "InteractiveWorkflowArena", "run_interactive_workflow_arena", "ExternalSideEffectArena", "run_external_side_effect_arena"]
 from ts_agl.arena.safe_write_arena import SafeWriteArena, run_safe_write_arena
 from ts_agl.arena.interactive_workflow_arena import InteractiveWorkflowArena, run_interactive_workflow_arena
+from ts_agl.arena.external_side_effect_arena import ExternalSideEffectArena, run_external_side_effect_arena